+++ arboretum/permissions/checks/test_org_collaborators.py
@@ -12,17 +12,15 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
-"""Github organization direct collaborators check."""
-
-import json
+"""Repository organization/owner collaborators check."""
 
 from compliance.check import ComplianceCheck
-from compliance.evidence import DAY, ReportEvidence, get_evidence_by_path
+from compliance.evidence import DAY, ReportEvidence, evidences
 from compliance.utils.data_parse import get_sha256_hash
 
 
 class OrgCollaboratorsCheck(ComplianceCheck):
-    """Checks for organization repo direct collaborators."""
+    """Checks for repository organization/owner collaborators."""
 
     @property
     def title(self):
@@ -31,7 +29,7 @@ class OrgCollaboratorsCheck(ComplianceCheck):
 
         :returns: the title of the checks
         """
-        return 'Organization repo direct collaborators check'
+        return 'Repository Organization/Owner Collaborators'
 
     @classmethod
     def setUpClass(cls):
@@ -39,10 +37,10 @@ class OrgCollaboratorsCheck(ComplianceCheck):
         cls.config.add_evidences(
             [
                 ReportEvidence(
-                    'org_direct_collaborators.md',
+                    'org_collaborators.md',
                     'permissions',
                     DAY,
-                    'Organization repo direct collaborators report.'
+                    'Repository organization collaborators report.'
                 )
             ]
         )
@@ -51,6 +49,8 @@ class OrgCollaboratorsCheck(ComplianceCheck):
     def test_org_direct_collaborators(self):
         """Check that there are no direct collaborators in the org repos."""
         orgs = self.config.get('org.permissions.org_integrity.orgs')
+        evidence_paths = {}
+        exceptions = {}
         for org in orgs:
             if 'direct' not in org.get('collaborator_types', []):
                 continue
@@ -62,48 +62,47 @@ class OrgCollaboratorsCheck(ComplianceCheck):
                 service = 'bb'
 
             url_hash = get_sha256_hash([org['url']], 10)
-            fname = f'{service}_direct_collaborators_{url_hash}.json'
-            repos = json.loads(
-                get_evidence_by_path(f'raw/permissions/{fname}',
-                                     self.locker).content
-            )
-            for repo in repos:
-                all_users = [u['login'] for u in repos[repo]]
+            filename = f'{service}_direct_collaborators_{url_hash}.json'
+            path = f'raw/permissions/{filename}'
+            evidence_paths[org_name] = path
+            exceptions[org_name] = org.get('exceptions', [])
+        with evidences(self, evidence_paths) as raws:
+            self._generate_results(raws, exceptions)
+
+    def _generate_results(self, evidences, exceptions):
+        for org, ev in evidences.items():
+            for repo in ev.content_as_json:
+                all_users = [u['login'] for u in ev.content_as_json[repo]]
                 if not all_users:
                     continue
-                exceptions = [
+                exception_users = [
                     e['user']
-                    for e in org.get('exceptions', [])
+                    for e in exceptions[org]
                     if ('repos' not in e or repo in e['repos'])
                 ]
-                failed_users = set(all_users) - set(exceptions)
-                warning_users = set(all_users).intersection(exceptions)
+                failed_users = set(all_users) - set(exception_users)
+                warning_users = set(all_users).intersection(exception_users)
                 if failed_users:
                     self.add_failures(
                         'unexpected-org-collaborators', {
-                            'org': org_name,
-                            'repo': repo,
-                            'users': failed_users
+                            'org': org, 'repo': repo, 'users': failed_users
                         }
                     )
                 if warning_users:
                     self.add_warnings(
                         'allowed-org-collaborators', {
-                            'org': org_name,
-                            'repo': repo,
-                            'users': warning_users
+                            'org': org, 'repo': repo, 'users': warning_users
                         }
                     )
 
-    def msg_org_direct_collaborators(self):
+    def get_notification_message(self):
         """
-        Organization repository direct collaborators notifier.
+        Repository organization collaborators check notifier.
 
         :returns: notification dictionary
         """
         return {
-            'subtitle': 'Unexpected direct collaborators in organizations',
-            'body': None
+            'subtitle': 'Repository organization collaborators', 'body': None
         }
 
     def get_reports(self):
@@ -112,4 +111,4 @@ class OrgCollaboratorsCheck(ComplianceCheck):
 
         :returns: the report(s) generated for this check
         """
-        return ['permissions/org_direct_collaborators.md']
+        return ['permissions/org_collaborators.md']