+++ CHANGES.md
@@ -1,6 +1,6 @@
 # [0.10.0](https://github.com/ComplianceAsCode/auditree-arboretum/releases/tag/v0.10.0)
 
-- [ADDED] Organization repository direct collaborators check added to `Permissions`
+- [ADDED] Organization repository direct collaborators check added to `permissions`.
 
 # [0.9.0](https://github.com/ComplianceAsCode/auditree-arboretum/releases/tag/v0.9.0)